feat(Snowfall)!: Abstract the snowflake rendering logic into the SnowfallCanvas class that is independent of React.
This allows the library to be used without React if that's your preference! You can actually use this with any library or even with vanilla JS.

BREAKING CHANGE: Lots of internals have moved around - if you are only using the `<Snowfall />` component then you shouldn't have to make any changes. If you were importing any of the library internals such as the custom hooks or some utilities then these have been reorganized and are now not exported from the root of the package - you will need to import them from the specific files.